+++ packages/non-deployable/token-allocation/test-payload.js
@@ -7,7 +7,6 @@ const environmentVariables = [
   { envKey: 'COINRANKING_ADAPTER_URL', value: 'coinranking' },
   { envKey: 'CRYPTOCOMPARE_ADAPTER_URL', value: 'cryptocompare' },
   { envKey: 'KAIKO_ADAPTER_URL', value: 'kaiko' },
-  { envKey: 'NOMICS_ADAPTER_URL', value: 'nomics' },
 ]
 
 function searchEnvironment(environmentVariables) {