+++ packages/scripts/src/get-changed-adapters/soakTestBlacklist.ts
@@ -30,7 +30,6 @@ export const SoakTestBlacklist: string[] = [
   'ethgasstation',
   'ethwrite',
   'expert-car-broker',
-  'finnhub',
   'flightaware',
   'fmpcloud',
   'genesis-volatility',