+++ packages/composites/market-closure/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/market-closure-adapter
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [201c92cc4]
+  - @chainlink/finnhub-adapter@2.1.0
+
 ## 1.3.0
 
 ### Minor Changes